작성자만 수정,삭제할 수 있게 그리고 관리자만 공지게시판 생성수정삭제 가능하게 만듬

+++ server/controller/Cerrands.js
@@ -13,6 +13,11 @@ exports.userLogin = async (req, res) => {
       },
     });
     if (result != null) {
+      console.log(result.dataValues.user_name);
+      req.session.user_name = result.dataValues.user_name;
+      req.session.user_type = result.dataValues.user_type;
+      console.log(req.session.user_name);
+      console.log(req.session.user_type);
       res.send(true);
     } else {
       res.send(false);
@@ -29,9 +34,24 @@ exports.userRegister = async (req, res) => {
       user_id: req.body.user_id,
       user_pw: req.body.user_pw,
       user_name: req.body.user_name,
-      user_type: req.body.user_type.value,
+      user_type: req.body.user_type,
     });
     res.end();
+  } catch (err) {
+    res.send(err);
+  }
+};
+
+// 로그아웃
+exports.userLogout = async (req, res) => {
+  try {
+    await req.session.destroy((err) => {
+      if (err) {
+        throw err;
+      }
+      req.session;
+      res.send(true);
+    });
   } catch (err) {
     res.send(err);
   }
@@ -85,7 +105,7 @@ exports.create_wanter_board = async (req, res) => {
       wanter_board_title: req.body.wanter_board_title,
       wanter_board_content: req.body.wanter_board_content,
       wanter_board_place: req.body.wanter_board_place,
-      wanter_board_dead_line: req.body.wanter_board_dead_line.value,
+      wanter_board_dead_line: req.body.wanter_board_dead_line,
       wanter_board_done: false,
     });
     res.send(result);
@@ -96,23 +116,29 @@ exports.create_wanter_board = async (req, res) => {
 
 // 게시물 수정
 exports.update_wanter_board = async (req, res) => {
-  console.log("**********************8");
-  console.log(req.params.boardId);
-  console.log(req.params.wanter_board_id);
   try {
-    const [result] = await Errands.Wanter_board.update(
-      {
-        wanter_board_title: req.body.wanter_board_title,
-        wanter_board_content: req.body.wanter_board_place,
-        wanter_board_place: req.body.wanter_board_place,
-        wanter_board_done: req.body.done,
-      },
-      { where: { wanter_board_id: { [Op.eq]: req.params.boardId } } }
-    );
-    if (result === 0) {
-      return res.send(false);
+    const auth = await Errands.User_info.findOne({
+      attributes: ["user_name"],
+      where: { user_name: { [Op.eq]: req.session.user_name } },
+    });
+    console.log(auth);
+    if (auth.dataValues.user_name == req.session.user_name) {
+      const [result] = Errands.Wanter_board.update(
+        {
+          wanter_board_title: req.body.wanter_board_title,
+          wanter_board_content: req.body.wanter_board_place,
+          wanter_board_place: req.body.wanter_board_place,
+          wanter_board_done: req.body.done,
+        },
+        { where: { wanter_board_id: { [Op.eq]: req.params.boardId } } }
+      );
+      if (result === 0) {
+        return res.send(false);
+      }
+      res.send(true);
+    } else {
+      res.send("작성자만 수정할 수 있습니다");
     }
-    res.send(true);
   } catch (err) {
     res.send(err);
   }
@@ -121,13 +147,21 @@ exports.update_wanter_board = async (req, res) => {
 // 게시물 삭제
 exports.delete_wanter_board = async (req, res) => {
   try {
-    const result = await Errands.Wanter_board.destroy({
-      where: { wanter_board_id: { [Op.eq]: req.params.boardId } },
+    const auth = await Errands.User_info.findOne({
+      attributes: ["user_name"],
+      where: { user_name: { [Op.eq]: req.session.user_name } },
     });
-    if (!result) {
-      return res.send(false);
+    if (auth.dataValues.user_name == req.session.user_name) {
+      const result = Errands.Wanter_board.destroy({
+        where: { wanter_board_id: { [Op.eq]: req.params.boardId } },
+      });
+      if (!result) {
+        return res.send(false);
+      }
+      res.send(true);
+    } else {
+      res.send("작성자만 삭제할 수 있습니다");
     }
-    res.send(true);
   } catch (err) {
     res.send(err);
   }
@@ -149,7 +183,7 @@ exports.create_wanter_comment = async (req, res) => {
   try {
     const result = await Errands.Wanter_comment.create({
       wanter_comment_board_id: req.params.boardId,
-      wanter_comment_writer: req.body.wanter_comment_writer,
+      wanter_comment_writer: req.session.user_name,
       wanter_comment_content: req.body.wanter_comment_content,
     });
     res.send(result);
@@ -161,21 +195,29 @@ exports.create_wanter_comment = async (req, res) => {
 // 댓글 수정
 exports.update_wanter_comment = async (req, res) => {
   try {
-    const [result] = await Errands.Wanter_comment.update(
-      {
-        wanter_comment_content: req.body.wanter_comment_content,
-      },
-      {
-        where: {
-          wanter_comment_id: { [Op.eq]: req.params.commentId },
-          wanter_comment_board_id: { [Op.eq]: req.params.boardId },
+    const auth = await Errands.User_info.findOne({
+      attributes: ["user_name"],
+      where: { user_name: { [Op.eq]: req.session.user_name } },
+    });
+    if (auth.dataValues.user_name == req.session.user_name) {
+      const [result] = Errands.Wanter_comment.update(
+        {
+          wanter_comment_content: req.body.wanter_comment_content,
         },
+        {
+          where: {
+            wanter_comment_id: { [Op.eq]: req.params.commentId },
+            wanter_comment_board_id: { [Op.eq]: req.params.boardId },
+          },
+        }
+      );
+      if (result === 0) {
+        return res.send(false);
       }
-    );
-    if (result === 0) {
-      return res.send(false);
+      res.send(true);
+    } else {
+      res.send("작성자만 삭제할 수 있습니다");
     }
-    res.send(true);
   } catch (err) {
     res.send(err);
   }
@@ -184,16 +226,24 @@ exports.update_wanter_comment = async (req, res) => {
 // 댓글 삭제
 exports.delete_wanter_comment = async (req, res) => {
   try {
-    const result = await Errands.Wanter_comment.destroy({
-      where: {
-        wanter_comment_id: { [Op.eq]: req.params.commentId },
-        wanter_comment_board_id: { [Op.eq]: req.params.boardId },
-      },
+    const auth = await Errands.User_info.findOne({
+      attributes: ["user_name"],
+      where: { user_name: { [Op.eq]: req.session.user_name } },
     });
-    if (!result) {
-      return res.send(false);
+    if (auth.dataValues.user_name == req.session.user_name) {
+      const result = Errands.Wanter_comment.destroy({
+        where: {
+          wanter_comment_id: { [Op.eq]: req.params.commentId },
+          wanter_comment_board_id: { [Op.eq]: req.params.boardId },
+        },
+      });
+      if (!result) {
+        return res.send(false);
+      }
+      res.send(true);
+    } else {
+      res.send("작성자만 댓글 삭제할 수 있습니다");
     }
-    res.send(true);
   } catch (err) {
     res.send(err);
   }
@@ -217,7 +267,7 @@ exports.read_few_helper_board = async (req, res) => {
 exports.read_helper_board = async (req, res) => {
   try {
     const result = await Errands.Helper_board.findAll({
-      helper_board_writer: req.body.helper_board_writer,
+      helper_board_writer: req.session.user_name,
       helper_board_title: req.body.helper_board_title,
       helper_board_content: req.body.helper_board_content,
       helper_board_place: req.body.helper_board_place,
@@ -245,7 +295,7 @@ exports.read_one_helper_board = async (req, res) => {
 exports.create_helper_board = async (req, res) => {
   try {
     const result = await Errands.Helper_board.create({
-      helper_board_writer: req.body.helper_board_writer,
+      helper_board_writer: req.session.user_name,
       helper_board_title: req.body.helper_board_title,
       helper_board_content: req.body.helper_board_content,
       helper_board_place: req.body.helper_board_place,
@@ -260,21 +310,29 @@ exports.create_helper_board = async (req, res) => {
 // 게시물 수정
 exports.update_helper_board = async (req, res) => {
   try {
-    const [result] = await Errands.Helper_board.update(
-      {
-        helper_board_title: req.body.helper_board_title,
-        helper_board_content: req.body.helper_board_content,
-        helper_board_place: req.body.helper_board_place,
-        helper_board_done: req.body.done,
-      },
-      {
-        where: { helper_board_id: { [Op.eq]: req.params.boardId } },
+    const auth = await Errands.User_info.findOne({
+      attributes: ["user_name"],
+      where: { user_name: { [Op.eq]: req.session.user_name } },
+    });
+    if (auth.dataValues.user_name == req.session.user_name) {
+      const [result] = Errands.Helper_board.update(
+        {
+          helper_board_title: req.body.helper_board_title,
+          helper_board_content: req.body.helper_board_content,
+          helper_board_place: req.body.helper_board_place,
+          helper_board_done: req.body.done,
+        },
+        {
+          where: { helper_board_id: { [Op.eq]: req.params.boardId } },
+        }
+      );
+      if (result === 0) {
+        return res.send(false);
       }
-    );
-    if (result === 0) {
-      return res.send(false);
+      res.send(true);
+    } else {
+      res.send("작성자만 수정할 수 있습니다");
     }
-    res.send(true);
   } catch (err) {
     res.send(err);
   }
@@ -283,13 +341,21 @@ exports.update_helper_board = async (req, res) => {
 // 게시물 삭제
 exports.delete_helper_board = async (req, res) => {
   try {
-    const result = await Errands.Helper_board.destroy({
-      where: { helper_board_id: { [Op.eq]: req.params.boardId } },
+    const auth = await Errands.User_info.findOne({
+      attributes: ["user_name"],
+      where: { user_name: { [Op.eq]: req.session.user_name } },
     });
-    if (!result) {
-      return res.send(false);
+    if (auth.dataValues.user_name == req.session.user_name) {
+      const result = Errands.Helper_board.destroy({
+        where: { helper_board_id: { [Op.eq]: req.params.boardId } },
+      });
+      if (!result) {
+        return res.send(false);
+      }
+      res.send(true);
+    } else {
+      res.send("작성자만 삭제할 수 있습니다");
     }
-    res.send(true);
   } catch (err) {
     res.send(err);
   }
@@ -313,7 +379,7 @@ exports.create_helper_comment = async (req, res) => {
   try {
     const result = await Errands.Helper_comment.create({
       helper_comment_board_id: req.params.boardId,
-      helper_comment_writer: req.body.helper_comment_writer,
+      helper_comment_writer: req.session.user_name,
       helper_comment_content: req.body.helper_comment_content,
     });
     res.send(result);
@@ -325,21 +391,29 @@ exports.create_helper_comment = async (req, res) => {
 // 댓글 수정
 exports.update_helper_comment = async (req, res) => {
   try {
-    const [result] = await Errands.Helper_comment.update(
-      {
-        helper_comment_content: req.body.helper_comment_content,
-      },
-      {
-        where: {
-          helper_comment_id: { [Op.eq]: req.params.commentId },
-          helper_comment_board_id: { [Op.eq]: req.params.boardId },
+    const auth = await Errands.User_info.findOne({
+      attributes: ["user_name"],
+      where: { user_name: { [Op.eq]: req.session.user_name } },
+    });
+    if (auth.dataValues.user_name == req.session.user_name) {
+      const [result] = Errands.Helper_comment.update(
+        {
+          helper_comment_content: req.body.helper_comment_content,
         },
+        {
+          where: {
+            helper_comment_id: { [Op.eq]: req.params.commentId },
+            helper_comment_board_id: { [Op.eq]: req.params.boardId },
+          },
+        }
+      );
+      if (result === 0) {
+        return res.send(false);
       }
-    );
-    if (result === 0) {
-      return res.send(false);
+      res.send(true);
+    } else {
+      res.send("작성자만 수정할 수 있습니다");
     }
-    res.send(true);
   } catch (err) {
     res.send(err);
   }
@@ -348,16 +422,24 @@ exports.update_helper_comment = async (req, res) => {
 // 댓글 삭제
 exports.delete_helper_comment = async (req, res) => {
   try {
-    const result = await Errands.Helper_comment.destroy({
-      where: {
-        helper_comment_id: { [Op.eq]: req.params.commentId },
-        helper_comment_board_id: { [Op.eq]: req.params.boardId },
-      },
+    const auth = await Errands.User_info.findOne({
+      attributes: ["user_name"],
+      where: { user_name: { [Op.eq]: req.session.user_name } },
     });
-    if (!result) {
-      return res.send(false);
+    if (auth.dataValues.user_name == req.session.user_name) {
+      const result = Errands.Helper_comment.destroy({
+        where: {
+          helper_comment_id: { [Op.eq]: req.params.commentId },
+          helper_comment_board_id: { [Op.eq]: req.params.boardId },
+        },
+      });
+      if (!result) {
+        return res.send(false);
+      }
+      res.send(true);
+    } else {
+      res.send("작성자만 삭제할 수 있습니다");
     }
-    res.send(true);
   } catch (err) {
     res.send(err);
   }
@@ -404,12 +486,21 @@ exports.read_one_notice = async (req, res) => {
 // 공지 생성
 exports.create_notice = async (req, res) => {
   try {
-    const result = await Errands.notice.create({
-      notice_writer: req.body.notice_writer,
-      notice_title: req.body.notice_title,
-      notice_content: req.body.notice_content,
+    const auth = await Errands.User_info.findOne({
+      attributes: ["user_type"],
+      where: { user_type: { [Op.eq]: req.session.user_type } },
     });
-    res.send(result);
+    if (auth.dataValues.user_type == "root") {
+      const result = Errands.Notice.create({
+        notice_writer: req.session.user_name,
+        notice_title: req.body.notice_title,
+        notice_content: req.body.notice_content,
+      });
+
+      res.send(result);
+    } else {
+      res.send("생성 권한이 없습니다");
+    }
   } catch (err) {
     res.send(err);
   }
@@ -418,17 +509,25 @@ exports.create_notice = async (req, res) => {
 // 공지 수정
 exports.update_notice = async (req, res) => {
   try {
-    const [result] = await Errands.Notice.update(
-      {
-        notice_title: req.body.notice_title,
-        notice_content: req.body.notice_content,
-      },
-      { where: { notice_id: { [Op.eq]: req.params.boardId } } }
-    );
-    if (result === 0) {
-      return res.send(false);
+    const auth = await Errands.User_info.findOne({
+      attributes: ["user_type"],
+      where: { user_type: { [Op.eq]: req.session.user_type } },
+    });
+    if (auth.dataValues.user_type == "root") {
+      const [result] = Errands.Notice.update(
+        {
+          notice_title: req.body.notice_title,
+          notice_content: req.body.notice_content,
+        },
+        { where: { notice_id: { [Op.eq]: req.params.boardId } } }
+      );
+      if (result === 0) {
+        return res.send(false);
+      }
+      res.send(true);
+    } else {
+      res.send("수정 권한이 없습니다");
     }
-    res.send(true);
   } catch (err) {
     res.send(err);
   }
@@ -437,13 +536,21 @@ exports.update_notice = async (req, res) => {
 // 공지 삭제
 exports.delete_notice = async (req, res) => {
   try {
-    const result = await Errands.Notice.destroy({
-      where: { notice_id: { [Op.eq]: req.params.boardId } },
+    const auth = await Errands.User_info.findOne({
+      attributes: ["user_type"],
+      where: { user_type: { [Op.eq]: req.session.user_type } },
     });
-    if (!result) {
-      return res.send(false);
+    if (auth.dataValues.user_type == "root") {
+      const result = Errands.Notice.destroy({
+        where: { notice_id: { [Op.eq]: req.params.boardId } },
+      });
+      if (!result) {
+        return res.send(false);
+      }
+      res.send(true);
+    } else {
+      res.send("삭제 권한이 없습니다");
     }
-    res.send(true);
   } catch (err) {
     res.send(err);
   }
